SearchInput 주석추가, 파일 분리
